+++ src/main/java/fsm/process/ChooseSyntax.java
@@ -1,8 +1,6 @@
 package fsm.process;
 
 import fsm.Guard;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 import java.util.ArrayList;
@@ -10,7 +8,6 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ChooseSyntax {
     final List<Option> options = new ArrayList<>();
 
@@ -18,7 +15,6 @@ public class ChooseSyntax {
         options.add(new Option(guard, config));
         return this;
     }
-
     ChooseSyntax otherwise(Consumer<Process> config) {
         Guard otherwise = Guard.and(
             options.stream()
@@ -27,7 +23,6 @@ public class ChooseSyntax {
         );
         return when(otherwise, config);
     }
-
     static ChooseSyntax choose() {
         return new ChooseSyntax();
     }
